+++ src/main/java/com/itb/inf3em/studyconnect/model/repository/MaterialRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MaterialRepository extends CrudRepository<Material,Integer> {
+public interface MaterialRepository extends CrudRepository<Material, Long> {
 }
